requirements.txt
chatterbox-tts==0.1.2
moviepy>=2.2.1
torchaudio>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0